Fix regex for nick highlighting
The highlighting regex was using [\W\D] which would match any non-word
char or any non-digit char. So single letter additions to a nick would
be highlighted, too.

We basically want it to match at a word boundary. Since the word
boundary is implemented as (^\w|\w$|\W\w|\w\W) that won't work with
Unicode nicks or any nicks using non word chars like `. So we can't use
\b but have to use only \W. That will still missmatch Unicode nicks as
part of longer Unicode sequences, but that's a limitation with regexes
in Javascript, not a regression.